Revert "Improve description formatting and add util test coverage"
This reverts commit 879757750745cf09d24d23b374087ce24ff9dac2.
