+++ packages/manager/src/features/Support/SupportTickets/ticketUtils.ts
@@ -1,20 +1,9 @@
-import { getTickets } from '@linode/api-v4/lib/support';
+import { Filter, Params } from '@linode/api-v4';
+import { TicketSeverity, getTickets } from '@linode/api-v4/lib/support';
 
 import { useAccountManagement } from 'src/hooks/useAccountManagement';
 import { useFlags } from 'src/hooks/useFlags';
 import { isFeatureEnabled } from 'src/utilities/accountCapabilities';
-
-import {
-  SMTP_FIELD_NAME_TO_LABEL_MAP,
-  TICKET_TYPE_TO_CUSTOM_FIELD_KEYS_MAP,
-} from './constants';
-
-import type {
-  AllSupportTicketFormFields,
-  SupportTicketFormFields,
-  TicketType,
-} from './SupportTicketDialog';
-import type { Filter, Params } from '@linode/api-v4';
 
 /**
  * getStatusFilter
@@ -71,40 +60,8 @@ export const useTicketSeverityCapability = () => {
   );
 };
 
-/**
- * formatDescription
- *
- * When variant ticketTypes include additional fields, fields must concat to one description string.
- * For readability, replace field names with field labels and format the description in Markdown.
- * @param values - the form payload, which can either be the general fields, or the general fields plus any custom fields
- * @param ticketType - either 'general' or a custom ticket type (e.g. 'smtp')
- *
- * @returns a description string containing custom fields in Markdown format
- */
-export const formatDescription = (
-  values: AllSupportTicketFormFields | SupportTicketFormFields,
-  ticketType: TicketType
-) => {
-  type customFieldTuple = [string, string | undefined];
-  const customFields: customFieldTuple[] = Object.entries(
-    values
-  ).filter(([key, _value]: customFieldTuple) =>
-    TICKET_TYPE_TO_CUSTOM_FIELD_KEYS_MAP[ticketType]?.includes(key)
-  );
-
-  // If there are no custom fields, just return the initial description.
-  if (customFields.length === 0) {
-    return values.description;
-  }
-
-  // Add all custom fields to the description in the ticket body, to be viewed on ticket details page and by Customer Support.
-  return customFields
-    .map(([key, value]) => {
-      let label = key;
-      if (ticketType === 'smtp') {
-        label = SMTP_FIELD_NAME_TO_LABEL_MAP[key];
-      }
-      return `**${label}**\n${value ? value : 'No response'}`;
-    })
-    .join('\n\n');
-};
+export const severityLabelMap: Map<TicketSeverity, string> = new Map([
+  [1, '1-Major Impact'],
+  [2, '2-Moderate Impact'],
+  [3, '3-Low Impact'],
+]);
